module/power_domain: Move contexts into new internal header file
The module and domain contexts are now in their own internal header.
Indicies for internal events and structs are also now in this header.
